+++ python/submit.py
@@ -70,7 +70,11 @@ class submit(gr.basic_block):
         self.request['timestamp'] = timestamp.isoformat()[:-3] + 'Z'
 
         params = bytes(urllib.parse.urlencode(self.request), encoding = 'ascii')
-        f = urllib.request.urlopen('{}?{}'.format(self.url, params), data=params)
+        try:
+            f = urllib.request.urlopen('{}?{}'.format(self.url, params), data=params)
+        except Exception as e:
+            print('Error while submitting telemetry:', e)
+            return
         reply = f.read()
         code = f.getcode()
         if code < 200 or code >= 300: